+++ lib/components/indicators/turnoverrate_indicator.dart
@@ -15,8 +15,8 @@ class TurnoverRateIndicator extends StatefulWidget {
   final double height;
   final List<UiKline> klines;
   final UiKlineRange klineRange;
-  final int klineWidth;
-  final int klineInnerWidth;
+  final double klineWidth;
+  final double klineInnerWidth;
   final int crossLineIndex;
   const TurnoverRateIndicator({
     super.key,
@@ -65,8 +65,8 @@ class _TurnoverRatePainter extends CustomPainter {
   final List<UiKline> klines;
   final UiKlineRange klineRng;
   final int crossLineIndex;
-  final int klineWidth;
-  final int klineInnerWidth;
+  final double klineWidth;
+  final double klineInnerWidth;
   final List<bool> isUpList;
 
   _TurnoverRatePainter({
@@ -146,8 +146,8 @@ class _TurnoverRatePainter extends CustomPainter {
 
     double maxTurnoverRate = _calcMaxTurnoverRate();
     for (int i = klineRng.begin; i < klineRng.end; i++) {
-      final x = i * klineWidth.toDouble();
-      final barWidth = klineInnerWidth.toDouble();
+      final x = i * klineWidth;
+      final barWidth = klineInnerWidth;
       final barHeight = (klines[i].turnoverRate / maxTurnoverRate) * bodyHeight;
       final y = titleHeight + bodyHeight - barHeight;
